refactor(focus-group): migrate to code mode architecture for better maintainability
- Restructure codebase into core, codemode, and mcp layers following architecture guidelines
- Separate business logic, visualization, and MCP protocol handling into dedicated modules
- Update package version to 0.4.0 to reflect architectural changes
- Add comprehensive test suite for the new code mode API
- Remove old monolithic implementation in favor of modular design
