+++ .trae/documents/plan_20260212_195249.md
@@ -1,0 +1,41 @@
+# Migrate Focus Group Server to Code Mode Architecture
+
+I have analyzed the current `focus-group` server and the Code Mode architecture guidelines. I propose the following migration plan to be executed using the `mcp-code-architect` agent.
+
+## 1. Architecture Design
+We will restructure `src/focus-group/src/` into the standard 3-layer Code Mode architecture:
+
+### A. Core Layer (`src/core/`)
+- **`types.ts`**: Will contain all domain interfaces (`FocusGroupPersona`, `Feedback`, `FocusAreaAnalysis`, `FocusGroupData`).
+- **`logic.ts`**: Will contain the `FocusGroupLogic` class (renamed from `FocusGroupServer`), handling validation, state updates, and business logic.
+- **`visualizer.ts`**: Will contain the `chalk`-based visualization logic to keep the core logic clean.
+
+### B. Code Mode Layer (`src/codemode/`)
+- **`index.ts`**: Will export a strongly-typed `FocusGroupServer` class (API).
+- This API will expose methods like `processFocusGroup(data: FocusGroupData)` and return structured results, wrapping the Core logic.
+
+### C. MCP Adapter Layer (`src/mcp/`)
+- **`tools.ts`**: Will define the `focusGroup` tool JSON schema.
+- **`server.ts`**: Will handle the MCP protocol connection and map the `focusGroup` tool to the Code Mode API.
+
+### D. Entry Point (`src/index.ts`)
+- Will export the Code Mode API for library usage.
+- Will conditionally start the MCP server when run directly (for `bunx` usage).
+
+## 2. Migration Steps
+1.  **Scaffold Directories**: Create `core`, `codemode`, and `mcp` directories.
+2.  **Refactor & Move**:
+    - Extract interfaces to `core/types.ts`.
+    - Move business logic and state management to `core/logic.ts`.
+    - Move visualization code to `core/visualizer.ts`.
+3.  **Implement Code Mode API**: Create the programmable API wrapper in `codemode/index.ts`.
+4.  **Implement MCP Adapter**: Create the tool definition and server runner in `mcp/`.
+5.  **Update Entry Point**: Rewrite `index.ts` to follow the pattern.
+6.  **Migrate Tests**: Refactor `src/index.test.ts` to test the new Code Mode API instead of the raw MCP server.
+
+## 3. Verification
+- **Compilation**: Ensure `tsc` (or `bun build`) runs without errors.
+- **Tests**: Run `bun test` to verify functionality is preserved.
+- **Manual Check**: Verify the `chalk` output still works as expected.
+
+I will use the `mcp-code-architect` agent to execute this transformation efficiently and accurately.
